changed scoring
changed score to double, bumped up number of games played during rollout to 200 and then divided wins by total games. seems to be playing a little better but rollout might need to be re done if someone has time

+++ team-00/src/main/java/ubc/cosc322/Agent.java
@@ -24,7 +24,7 @@ public class Agent extends GamePlayer{
 	
 	private Queen currentQueen = null;
 	private ArrayList<Integer> nextMove = null;
-	private int score = 0;
+	private double score = 0;
 	
 	String ourAmazon;
 	String otherAmazon;
@@ -78,6 +78,7 @@ public class Agent extends GamePlayer{
     	    			this.isWhite = true;
     	    			board.setAIColor(true);
     	    			queens = setTeam(board);
+
     	    		} else {
     	    			System.out.println("The AI is black");
     	    			this.isWhite = false;
